separate array and buffer tests
